refactor: split the presentation layer
