+++ src/app.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { UserController } from './user.controller';
+import { UserController } from './presentation/user.controller';
 import { UserModule } from './application/user.module';
 
 @Module({
